fix clippy basic

+++ jolt-core/src/jolt/instruction/addi.rs
@@ -56,7 +56,7 @@ impl<const WORD_SIZE: usize> LookupQuery<WORD_SIZE> for RISCVCycle<ADDI> {
             #[cfg(test)]
             8 => (x as u8).overflowing_add(y as u8).0 as u64,
             32 => (x as u32).overflowing_add(y as u32).0 as u64,
-            64 => (x as u64).overflowing_add(y as u64).0,
+            64 => x.overflowing_add(y as u64).0,
             _ => panic!("{WORD_SIZE}-bit word size is unsupported"),
         }
     }
