+++ jolt-core/src/jolt/instruction/virtual_srai.rs
@@ -34,7 +34,7 @@ impl<const WORD_SIZE: usize> LookupQuery<WORD_SIZE> for RISCVCycle<VirtualSRAI> 
     fn to_lookup_output(&self) -> u64 {
         use crate::subprotocols::sparse_dense_shout::LookupBits;
         let (x, y) = LookupQuery::<WORD_SIZE>::to_instruction_inputs(self);
-        let mut x = LookupBits::new(x as u64, WORD_SIZE);
+        let mut x = LookupBits::new(x, WORD_SIZE);
         let mut y = LookupBits::new(y as u64, WORD_SIZE);
 
         let sign_bit = if x.leading_ones() == 0 { 0 } else { 1 };
